% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_helpers.r
\name{approx_rate}
\alias{approx_rate}
\title{Approximate instantaneous usage rate from a meter interpolation function}
\usage{
approx_rate(meter_fun, x, h = 1e-06)
}
\arguments{
\item{meter_fun}{Function created by \code{create_meter_fun}}

\item{x}{Time points at which the rate should be evaluated}

\item{h}{Step size for the numerical derivative}
}
\value{
Estimated usage rate at \code{x}
}
\description{
Approximate instantaneous usage rate from a meter interpolation function
}
