% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_helpers.r
\name{capture_plot}
\alias{capture_plot}
\title{Capture base plot output as a recordable object}
\usage{
capture_plot(expr)
}
\arguments{
\item{expr}{Expression containing plotting code}
}
\value{
\code{recordedplot} object
}
\description{
Evaluates an expression that generates a base R plot and
returns the plot as a \code{recordedplot} object.
}
