% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_helpers.r
\name{create_meter_fun}
\alias{create_meter_fun}
\title{Create an interpolation function for gas meter readings}
\usage{
create_meter_fun(dates, values)
}
\arguments{
\item{dates}{Vector of POSIXct timestamps}

\item{values}{Vector of cumulative meter readings corresponding to \code{dates}}
}
\value{
A function that can be used to interpolate the meter reading
}
\description{
Given dates and corresponding cumulative meter values, this function
creates a linear interpolation function to estimate the meter
at arbitrary time points.
}
