% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_helpers.r
\name{get_avg_temp}
\alias{get_avg_temp}
\title{Average temperature between two timestamps}
\usage{
get_avg_temp(df, var = "tavg", var_time = "Date", xmin, xmax)
}
\arguments{
\item{df}{Data frame containing at least a time column and a temperature column}

\item{var}{Name of the temperature column}

\item{var_time}{Name of the time column}

\item{xmin, xmax}{Start and end times for the interval}
}
\value{
Numeric average temperature over the interval
}
\description{
Average temperature between two timestamps
}
