% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_t_sum.r
\name{Maketsum}
\alias{Maketsum}
\title{Vectorized Version of maketsum}
\usage{
Maketsum(date, datelag, obs_hours.)
}
\arguments{
\item{date}{The end date of the interval for which the temperature sum is to be
calculated. This should be an object of class \code{Date}.}

\item{datelag}{The start date of the interval. This should also be an object of
class \code{Date}.}
}
\value{
A vector of normalized temperature sums for each pair of \code{date} and
\code{datelag} values.
}
\description{
The \code{Maketsum} function is a vectorized version of \code{maketsum}, allowing it to
be applied over vectors of dates and datelags.
}
\examples{
# Vectorizing over multiple intervals
dates <- as.Date(c("2023-01-31", "2023-02-28"))
datelags <- as.Date(c("2023-01-01", "2023-02-01"))
temp_sums <- Maketsum(dates, datelags)

}
