% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_weather.r
\name{simulate_weather}
\alias{simulate_weather}
\title{Simulate Weather Data}
\usage{
simulate_weather(
  n = 365,
  coefs. = list(tavg_acf = 0.833, tavg_sigma = 3.73, range_acf = 0.457, range_sigma =
    3.35)
)
}
\arguments{
\item{n}{An integer specifying the number of days for which to simulate weather
data. Default is 365, representing a year.}

\item{coefs}{A list containing the autocorrelation coefficients (acf) and standard
deviations (sigma) for the average temperature (\code{tavg_acf}, \code{tavg_sigma})
and temperature range (\code{range_acf}, \code{range_sigma}). Default values are
set for a typical year.}
}
\value{
A dataframe containing the simulated daily weather data, including day in
year, simulated year, predicted average temperature (\code{pred_tavg}), and
predicted temperature range (\code{pred_range}), as well as minimum (\code{tmin})
and maximum (\code{tmax}) temperatures for each day.
}
\description{
This function generates simulated weather data for a given number of days. It
specifically simulates daily average temperature and temperature range, accounting
for autocorrelation and variability in the data. The function is capable of
handling large simulations by segmenting the process.
}
\examples{
# Simulate weather data for a default one-year period
weather_data <- simulate_weather()

# Simulate weather data for 100 days with custom coefficients
custom_coefs <- list(tavg_acf = 0.8, tavg_sigma = 3.5, range_acf = 0.4, range_sigma = 3.0)
weather_data_custom <- simulate_weather(n = 100, coefs. = custom_coefs)

}
