% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yday_inverse.r
\name{yday_inverse}
\alias{yday_inverse}
\title{Convert Day of Year to Date}
\usage{
yday_inverse(day_in_year, yr = 2023)
}
\arguments{
\item{day_in_year}{An integer or character representing the day of the year.
For example, 1 for January 1st, 32 for February 1st, etc.}

\item{yr}{The year to which the day belongs, specified as an integer.
Default is 2023.}
}
\value{
A \code{Date} object representing the corresponding date in the specified year.
}
\description{
This function converts a given day of the year to a date object. It's useful for
converting numeric day-in-year representations back to standard date formats.
}
\examples{
# Convert the 100th day of 2023 to a date
date_from_day <- yday_inverse(100)

# Convert the 50th day of 2022 to a date
date_from_day_2022 <- yday_inverse(50, 2022)

}
